Upgrade setup-coron

+++ setup-coron/initbuild.sh
@@ -4,7 +4,6 @@ clear
 username=`whoami`
 thisDir=`pwd`
 
-
 addRulesFunc(){
 	read mIdVendor mIdProduct
 	echo "SUBSYSTEM==\"usb\", ATTR{idVendor}==\""$mIdVendor"\", ATTR{idProduct}==\""$mIdProduct"\", MODE=\"0600\"" | sudo tee -a /etc/udev/rules.d/51-android.rules
@@ -53,20 +52,18 @@ addRules(){
 
 addadbrules(){
 	echo -e "\n配置adb环境变量..."
-	cp 51-android.rules /etc/udev/rules.d/
-	chmod a+rx /etc/udev/rules.d/51-android.rules
-	source /etc/udev/rules.d/51-android.rules
+	sudo cp 51-android.rules /etc/udev/rules.d/
+	sudo chmod a+rx /etc/udev/rules.d/51-android.rules
 	echo "export PATH=$PATH:~/bin/" | sudo tee -a /etc/profile
 	source /etc/profile
 	sudo adb kill-server
 	sudo adb devices
 	echo "\n配置环境完成"
-	echo -e "按回车键继续"
-	read anykey
+	read -p "按回车键继续..."
 }
 
 changecoronlanguage(){
-echo -e "请输入coron项目所在目录(可以尝试把目录拖进来,如果失败请手动输入路径)"
+echo -e "请输入coron项目所在目录(可以把目录拖进来,)"
 read coronDir
 echo -e "输入1即可把coron项目环境改成中文"
 echo -e "输入2即可把coron项目环境改回英文"
@@ -77,15 +74,13 @@ case $languagechoose in
 		cd $coronDir
 		patch -p1<$thisDir/coron.patch
 		cd $thisDir
-		echo -e "按回车键继续"
-		read anykey
+		read -p "按回车键继续..."
 	;;
 	2)
 		cd $coronDir
 		patch -R -p1<$thisDir/coron.patch
 		cd $thisDir
-		echo -e "按回车键继续"
-		read anykey
+		read -p "按回车键继续..."
 	;;
 	*)
 		main
@@ -95,7 +90,7 @@ esac
 
 zipcenop(){
 	echo -e "这是刷机包或者apk&jar伪加密工具"
-	echo -e "输入你想加密的刷机包或者apk&jar(可以拖进来)"
+	echo -e "请把需要加密的刷机包或者apk&jar拖进来"
 	read cenopfile
 	echo -ne "\n选择:"
 	echo -e "输入1加密，输入2解密，输入任意字符退出"
@@ -103,14 +98,12 @@ zipcenop(){
 	read cenopmode
 case $cenopmode in
 	1)
-		java -jar $thisDir/ZipCenOp.jar r $cenopfile
-	echo -e "按回车键继续"
-	read anykey
+		java -jar $thisDir/ZipCenOp.jar e ${cenopfile//\'//}
+		read -p "按回车键继续..."
 	;;
 	2)
-		java -jar $thisDir/ZipCenOp.jar e $cenopfile
-	echo -e "按回车键继续"
-	read anykey
+		java -jar $thisDir/ZipCenOp.jar r ${cenopfile//\'//}
+		read -p "按回车键继续..."
 	;;
 	*)
 		main
@@ -135,8 +128,7 @@ repoSource(){
 	repo init -u https://github.com/baidurom/manifest.git -b "coron-"$($version)
 	repo sync
 	cd $thisDir
-	echo -e "按回车键继续"
-	read anykey
+	read -p "按回车键继续..."
 }
 
 fastrepoSource(){
@@ -156,8 +148,51 @@ fastrepoSource(){
 	repo init --repo-url git://github.com/baidurom/repo.git -u https://github.com/baidurom/manifest.git -b "coron-"$($version) --no-repo-verify
 	repo sync -c --no-clone-bundle --no-tags -j4
 	cd $thisDir
-	echo -e "按回车键继续"
-	read anykey
+	read -p "按回车键继续..."
+}
+
+installsdk(){
+echo
+echo "下载和配置 Android SDK!!"
+echo "请确保 unzip 已经安装"
+echo
+sudo apt-get install unzip -y
+if [ `getconf LONG_BIT` = "64" ];then
+	echo
+	echo "正在下载 Linux 64位 系统的Android SDK"
+	wget http://dl.google.com/android/adt/adt-bundle-linux-x86_64-20140702.zip
+	echo "下载完成!!"
+	echo "展开文件"
+	mkdir ~/adt-bundle
+	mv adt-bundle-linux-x86_64-20140702.zip ~/adt-bundle/adt_x64.zip
+	cd ~/adt-bundle
+	unzip adt_x64.zip
+	mv -f adt-bundle-linux-x86_64-20140702/* .
+	echo "正在配置"
+	echo -e '\n# Android tools\nexport PATH=${PATH}:~/adt-bundle/sdk/tools\nexport PATH=${PATH}:~/adt-bundle/sdk/platform-tools\nexport PATH=${PATH}:~/bin' >> ~/.bashrc
+	echo -e '\nPATH="$HOME/adt-bundle/sdk/tools:$HOME/adt-bundle/sdk/platform-tools:$PATH"' >> ~/.profile
+	echo "完成!!"
+else
+	echo
+	echo "正在下载 Linux 32位 系统的Android SDK"
+	wget http://dl.google.com/android/adt/adt-bundle-linux-x86-20140702.zip
+	echo "下载完成!!"
+	echo "展开文件"
+	mkdir ~/adt-bundle
+	mv adt-bundle-linux-x86-20140702.zip ~/adt-bundle/adt_x86.zip
+	cd ~/adt-bundle
+	unzip adt_x86.zip
+	mv -f adt-bundle-linux-x86_64-20140702/* .
+	echo "正在配置"
+	echo -e '\n# Android tools\nexport PATH=${PATH}:~/adt-bundle/sdk/tools\nexport PATH=${PATH}:~/adt-bundle/sdk/platform-tools\nexport PATH=${PATH}:~/bin' >> ~/.bashrc
+	echo -e '\nPATH="$HOME/adt-bundle/sdk/tools:$HOME/adt-bundle/sdk/platform-tools:$PATH"' >> ~/.profile
+	echo "完成!!"
+fi
+rm -Rf ~/adt-bundle/adt-bundle-linux-x86_64-20140702
+rm -Rf ~/adt-bundle/adt-bundle-linux-x86-20140702
+rm -f ~/adt-bundle/adt_x64.zip
+rm -f ~/adt-bundle/adt_x86.zip
+read -p "按回车键继续..."
 }
 
 initSystemConfigure(){
@@ -167,9 +202,10 @@ echo -e "\t1.ia32运行库"
 echo -e "\t2.JavaSE(Oracle Java JDK)"
 echo -e "\t3.aosp&cm&recovery编译环境"
 echo -e "\t4.adb运行环境"
-echo -e "\t5.coron项目中文环境"
-echo -e "\t6.hosts环境"
-echo -e "\t7.安卓开发必备环境(上面的1.2.3.4)"
+echo -e "\t5.AndroidSDK运行环境"
+echo -e "\t6.coron项目中文环境"
+echo -e "\t7.hosts环境"
+echo -e "\t8.安卓开发必备环境(上面12345）"
 echo -ne "\n选择:"
 read configurechoose
 case $configurechoose in
@@ -181,20 +217,19 @@ case $configurechoose in
 		echo -en "选择:"
 		read kind
 		if [ "$kind" == "1" ]; then
-			apt-get install ia32-libs
+			sudo apt-get install ia32-libs
 		else
 #start
 		cd /etc/apt/sources.list.d #进入apt源列表
-		echo "deb http://mirrors.oschina.net/ubuntu/ rarin main restricted universe multiverse" > ia32-libs-raring.list 
+		echo "deb http://old-releases.ubuntu.com/ubuntu/ raring main restricted universe multiverse" | sudo tee ia32-libs-raring.list
 #添加ubuntu 13.04的源，因为13.10的后续版本废弃了ia32-libs
 		sudo apt-get update #更新一下源
 		sudo apt-get install ia32-libs #安装ia32-libs
-		rm ia32-libs-raring.list #恢复源
+		sudo rm ia32-libs-raring.list #恢复源
 		sudo apt-get update #再次更新下源
 #end
 		fi
-		echo -e "按回车键继续"
-		read anykey
+		read -p "按回车键继续..."
 	;;
 	2)
 		sudo apt-get update
@@ -204,39 +239,57 @@ case $configurechoose in
 		echo -e "\n开始安装oracle java developement kit..."
 		sleep 1
 		sudo add-apt-repository ppa:webupd8team/java
-		sudo apt-get update && sudo apt-get install oracle-java6-installer
-		echo -e "按回车键继续"
-		read anykey
+		sudo apt-get update && sudo apt-get install oracle-java7-installer 
+		read -p "按回车键继续..."
 	;;
 	3)
 		echo -e "\n开始安装ROM编译环境..."
 		echo -e "请选择使用的系统版本:"
 		echo -e "\t1. ubuntu 12.04 及以下"
 		echo -e "\t2. 其他(包括deepin等基于ubuntu 的系统)"
+		echo -e "\t3. javaSE1.6.0"
 		echo -en "选择:"
 		read kind
 		if [ "$kind" == "1" ]; then
 			sudo apt-get install bison libc6 build-essential curl flex g++-multilib g++ gcc-multilib git-core gnupg gperf libesd0-dev libncurses5-dev libwxgtk2.8-dev lzop squashfs-tools xsltproc pngcrush schedtool zip zlib1g-dev 
+		elif [ "$kind" == "2" ]; then
+			sudo apt-get install bison ccache libc6 build-essential curl flex g++-multilib g++ gcc-multilib git-core gnupg gperf x11proto-core-dev tofrodos libx11-dev:i386 libgl1-mesa-dev libreadline6-dev:i386 libgl1-mesa-glx:i386 lib32ncurses5-dev libncurses5-dev:i386 lib32readLine-gplv2-dev lib32z1-dev libesd0-dev libncurses5-dev libsdl1.2-dev libwxgtk2.8-dev python-markdown libxml2 libxml2-utils lzop squashfs-tools xsltproc pngcrush schedtool zip zlib1g-dev:i386 zlib1g-dev	
+			sudo ln -s /usr/lib/i386-linux-gnu/mesa/libGL.so.1 /usr/lib/i386-linux-gnu/libGL.so
+		elif [ "$kind" == "3" ]; then
+			sudo apt-get update
+			echo -e "\n删除自带的openjdk..."
+			sleep 1
+			sudo apt-get purge openjdk-* icedtea-* icedtea6-*
+			echo -e "\n开始安装oracle java developement kit..."
+			sleep 1
+			sudo add-apt-repository ppa:webupd8team/java
+			sudo apt-get update && sudo apt-get install oracle-java6-installer 
+			echo "alias java-switch='sudo update-alternatives --config java'" | sudo tee -a /etc/profile
+			source /etc/profile
+			echo -e "你可以使用java-switch命令来切换java版本"
 		else
-			sudo apt-get install bison libc6 build-essential curl flex g++-multilib g++ gcc-multilib git-core gnupg gperf lib32ncurses5-dev lib32readLine-gplv2-dev lib32z1-dev libesd0-dev libncurses5-dev libsdl1.2-dev libwxgtk2.8-dev libxml2 libxml2-utils lzop squashfs-tools xsltproc pngcrush schedtool zip zlib1g-dev	
+			initSystemConfigure
 		fi
-		read anykey
+		read -p "按回车键继续..."
+	
 	;;
 	4)
+		sudo apt-get update
 		sudo apt-get install android-tools-adb android-tools-fastboot
 		addadbrules
-		echo -e "按回车键继续"
-		read anykey
+		read -p "按回车键继续..."
 	;;
 	5)
-		changecoronlanguage
+		installsdk
 	;;
 	6)
-		addhosts
-		echo -e "按回车键继续"
-		read anykey
+		changecoronlanguage
 	;;
 	7)
+		addhosts
+		read -p "按回车键继续..."
+	;;
+	8)
 		echo -e "\n开始配置32位运行环境..."
 		echo -e "请选择使用的系统版本:"
 		echo -e "\t1. ubuntu 12.04 及以下"
@@ -244,15 +297,15 @@ case $configurechoose in
 		echo -en "选择:"
 		read kind
 		if [ $kind -eq 1 ]; then
-			apt-get install ia32-libs
+			sudo apt-get install ia32-libs
 		else
 #start
 		cd /etc/apt/sources.list.d #进入apt源列表
-		echo "deb http://mirrors.oschina.net/ubuntu/ rarin main restricted universe multiverse" > ia32-libs-raring.list 
+		echo "deb http://old-releases.ubuntu.com/ubuntu/ raring main restricted universe multiverse" | sudo tee ia32-libs-raring.list
 #添加ubuntu 13.04的源，因为13.10的后续版本废弃了ia32-libs
 		sudo apt-get update #更新一下源
 		sudo apt-get install ia32-libs #安装ia32-libs
-		rm ia32-libs-raring.list #恢复源
+		sudo rm ia32-libs-raring.list #恢复源
 		sudo apt-get update #再次更新下源
 #end
 		fi
@@ -263,21 +316,21 @@ case $configurechoose in
 		echo -e "\n开始安装oracle java developement kit..."
 		sleep 1
 		sudo add-apt-repository ppa:webupd8team/java
-		sudo apt-get update && sudo apt-get install oracle-java6-installer
+		sudo apt-get update && sudo apt-get install oracle-java7-installer
 		sudo apt-get install android-tools-adb android-tools-fastboot
+		echo -e "\n开始安装AndroidSDK环境..."
+		installsdk
 		echo -e "\n开始安装ROM编译环境..."
 		sudo apt-get install bison build-essential curl flex g++-multilib g++ gcc-multilib git-core gnupg gperf lib32ncurses5-dev lib32readLine-gplv2-dev lib32z1-dev libesd0-dev libncurses5-dev libsdl1.2-dev libwxgtk2.8-dev libxml2 libxml2-utils lzop squashfs-tools xsltproc pngcrush schedtool zip zlib1g-dev
 		echo -e "\n配置adb环境变量..."
-		cp 51-android.rules /etc/udev/rules.d/
+		sudo cp 51-android.rules /etc/udev/rules.d/
 		sudo chmod a+rx /etc/udev/rules.d/51-android.rules
-		source /etc/udev/rules.d/51-android.rules
 		echo "export PATH=$PATH:~/bin/" | sudo tee -a /etc/profile
 		source /etc/profile
 		sudo adb kill-server
 		sudo adb devices
 		echo "\n配置环境完成"
-		echo -e "按回车键继续"
-		read anykey
+		read -p "按回车键继续..."
 	;;
 esac
 }
@@ -290,12 +343,12 @@ echo -ne "\n选择:"
 read logcatmode
 case $logcatmode in
 	1)
-		adb logcat > $thisDir/log
+		adb logcat -b main -b system -b radio > $thisDir/log
 	;;
 	2)
 		echo -e "输入你想过滤的内容"
 		read ignoretext
-		adb logcat|grep $ignoretext|tee $thisDir/log
+		adb logcat -b main -b system -b radio |grep $ignoretext|tee $thisDir/log
 	;;
 	*)
 		echo -e "请输入正确的命令"
@@ -308,23 +361,23 @@ clean(){
 	echo -e "正在清理残留文件"
 	rm -rf log
 	rm -rf screenshot.png
-	echo -e "按回车键退出"
-	read anykey
+	read -p "按回车键继续..."
 }
  
 main(){
 clear 
-echo -e "                 Android Development Shell scirpt"
-echo "          --made by Modificator & Edit by Jay_Kwan"
+echo -e "Android开发环境一键搭载脚本及开发工具"
+echo "--作者：Modificator & 嘉豪仔_Kwan"
 echo -e "			输入命令号码 :\n"
 echo -e "\t\t1. 使用root权限启动adb"
 echo -e "\t\t2. 设置环境变量"
-echo -e "\t\t3. 依然无法识别手机？没关系，选这个"
-echo -e "\t\t4. 同步源码"
-echo -e "\t\t5. 快速同步源码(跳过谷歌认证)"
-echo -e "\t\t6. 伪加密工具"
-echo -e "\t\t7. 抓取log工具"
-echo -e "\t\t8. 手机截图"
+echo -e "\t\t3. 安装安卓厨房（Android-Kitchen)"
+echo -e "\t\t4. 依然无法识别手机？没关系，选这个"
+echo -e "\t\t5. 同步源码"
+echo -e "\t\t6. 快速同步源码(跳过谷歌认证)"
+echo -e "\t\t7. 伪加密工具"
+echo -e "\t\t8. 抓取log工具"
+echo -e "\t\t9. 手机截图"
 echo -e "\t\t0. 离开脚本"
 echo -ne "\n选择:"
 read inp
@@ -332,8 +385,7 @@ case $inp in
 	1)
 		sudo adb kill-server
 		sudo adb devices
-		echo -e "按回车键继续"
-		read anykey
+		read -p "按回车键继续..."
 		main
 	;;
 	2)
@@ -341,34 +393,42 @@ case $inp in
 		main
 	;;
 	3)
-		addRules
+		echo "安装安卓厨房"
+		sudo apt-get install git -y
+		cd ~/
+		git clone https://github.com/kuairom/Android_Kitchen_cn
+		echo "安卓厨房已下载到主文件夹的Android_Kitchen_cn目录里！"
+		read -p "按回车键继续..."
 		main
 	;;
 	4)
-		repoSource
+		addRules
 		main
 	;;
 	5)
-		fastrepoSource
+		repoSource
 		main
 	;;
 	6)
-		zipcenop
+		fastrepoSource
 		main
 	;;
 	7)
-		logcat
+		zipcenop
 		main
 	;;
 	8)
+		logcat
+		main
+	;;
+	9)
 		adb shell /system/bin/screencap -p /data/local/tmp/screenshot.png
 		cd $thisDir
 		adb pull /data/local/tmp/screenshot.png
 		if [ "$?" == "0" ]; then
 		echo -e "截图文件已经输出到$thisDir"
 		fi
-		echo -e "按回车键继续"
-		read anykey
+		read -p "按回车键继续..."
 		main
 	;;
 	0)
@@ -378,7 +438,7 @@ case $inp in
 		echo -e "输入c清理残留文件否则直接退出"
 		echo -ne "\n输入c清理或者按回车退出:"
 		read cleanchoose
-		if [ "$cleanchoose" -eq "c" ]; then
+		if [ "$cleanchoose" == "c" ]; then
 			clean
 		fi
 		echo -e "\e[0m"
@@ -388,11 +448,14 @@ case $inp in
 	;;
 esac
 }
-echo -e "正在解压工具，请稍候......"
-if [ ! -f repo ]; then 
-	tar -xvf tools.tar
-fi 
-
+echo -e "正在检测，请稍候......"
+	if [ ! -f repo ]; then
+		echo -e "正在解压工具，请稍候......"
+		tar -xvf tools.tar
+	else
+		echo -e "工具已存在，跳过解压......"
+	fi
 sleep 1
+echo -e "说明：本脚本仅适用于Ubuntu及各大Ubuntu发行版使用，并且建议在14.04Lts版本下使用"
+read -p "按回车键继续..."
 main
-echo -e "\e[0m"
